Fix int32 overflow in Metal conv_general output offset for large tensors
The implicit_gemm_conv_2d_general Metal kernel computed output write
offsets using 32-bit int arithmetic. When batch * H_out * W_out * C_out
exceeded 2^31, the offset silently wrapped and corrupted tail batch
entries (zeroed out or wrote to wrong locations).

Promote offset_cm and offset to size_t in the output store loop,
matching the already-64-bit out_strides in MLXConvParams.

Add a regression test that crosses the 2^31 boundary with varying
per-batch input values so wrong-batch writes are also caught.

Fixes #3248

+++ mlx/backend/metal/kernels/steel/conv/kernels/steel_conv_general.h
@@ -200,14 +200,19 @@ implicit_gemm_conv_2d_general(
           (hw % jump_params->adj_out_w) * jump_params->f_out_jump_w + base_ow;
 
       if (n < params->N && oh < params->oS[0] && ow < params->oS[1]) {
-        int offset_cm = n * params->out_strides[0] +
-            oh * params->out_strides[1] + ow * params->out_strides[2];
+        size_t offset_cm =
+            static_cast<size_t>(n) *
+                static_cast<size_t>(params->out_strides[0]) +
+            static_cast<size_t>(oh) *
+                static_cast<size_t>(params->out_strides[1]) +
+            static_cast<size_t>(ow) *
+                static_cast<size_t>(params->out_strides[2]);
 
         STEEL_PRAGMA_UNROLL
         for (int j = 0; j < mma_t::TN; j++) {
           // Get accumulated result and associated offset in C
           thread const auto& accum = mma_op.Ctile.frag_at(i, j);
-          int offset = offset_cm + (j * mma_t::TN_stride);
+          size_t offset = offset_cm + (j * mma_t::TN_stride);
 
           constexpr short kelems = decltype(mma_op.Ctile)::kElemsPerFrag;
